Modify: utils folder created, relocate login_required, validation
